Initial implementation
This evolves the POC from the initial commit to an initial, somewhat
usable, implementation of ghasum.

This initial version supports an initialization command to start using
ghasum in a project. A verification command to check stored checksums of
a project against computed checksums for the actions used. And an update
command to update the stored checksums of a project with newly computed
checksums.