Add support for validating one workflow
Extend the `ghasum verify` command with the ability to verify a single
workflow. In particular, when the provided target is a file instead of a
directory it will treat it as if it's a workflow file inside a repo and
use the repo as target and limit the verification scope to only that
workflow.

The new behavior is tested primarily through `testscript` tests covering
both good weather and bad weather scenarios (this also improves the bad
weather scenarios for verification by 1. mixing valid and invalid
checksums for the "Checksum mismatch" scenario, which is a more useful
scenario; and 2. avoiding invalid checksums in the "Checksum missing"
scenario to make it more focussed). Additionally, local functionality
is unit tested in places where unit tests already exist.