Unlock gha.sum if an error occurs during updating
Fix a bug where using `ghasum update` could result in a locked gha.sum
file when an unexpected error occurred. While not fatal - it could be
fixed by re-initializing or successfully updating - it is inconvenient
and buggy. This fixes the problem by always unlocking the file in the
deferred function that closes the file handle as well.