Expand error test suite for `ghasum init`
Add three new error scenario tests for the `ghasum init` command. The
aim of this is to improve the program when things don't go as expected.

The first test corresponds to a scenario where there is a syntax error
in one of the workflows. In particular, this tests that there is a
failure and that the error makes sense. This test uncovered a problem in
the error message, namely that it didn't include the file where the
error occurred, which has been fixed as well.

The second tests corresponds to trying to initialize ghasum for a
project that does not use GitHub Actions (similar tests already exist
for updating and verifying). This is primarily a sanity check that the
CLI doesn't do anything in this scenario.

The third test is for a scenario where the target does not exist, in
which case the error should clearly state that that fact.

Additionally, this also corrects the initialized repository so that it
is valid, avoiding the command somehow failing because the sumfile is
invalid instead of it being initialized already.