Expand error test suite for `ghasum update`
Add two error scenario tests for the `ghasum update` command. The aim
of this is to improve the program when things don't go as expected.

The first test corresponds to a scenario where there is a syntax error
in one of the workflows. In particular, this tests that there is a
failure and that the error makes sense.

The second test is for a scenario where the target does not exist, in
which case the error should clearly state that that fact. This uncovered
a problem in the error message, namely that it just stated that the repo
has not been initialized rather than that there is no project.

Additionally, this aligns the test names and ordering with the error
test suite for `ghasum init` and `ghasum verify`.