Expand error test suite for `ghasum verify`
Add two error scenario tests for the `ghasum verify` command. The aim
of this is to improve the program when things don't go as expected.

The first test corresponds to a scenario where there is a syntax error
in the ghasum sumfile. In particular, this tests that there is a failure
and that the error makes sense.

The second test corresponds to a scenario where there is a syntax error
in one of the workflows. In particular, this tests that there is a
failure and that the error makes sense.

Additionally, this aligns the test names with the error test suite for
`ghasum init` and `ghasum verify`. It also aligns the test data within
the test suite.