Add support for validating one job in a workflow
Extend the `ghasum verify` command with the ability to verify a single
job in a workflow. In particular, when the provided target ends with
`:suffix` the suffix will be stripped and used as the job id/key, the
rest of the target is then treated as the workflow to verify. This will
result in the verification scope being limited to only that job (in that
workflow).

The new behavior is tested primarily through `testscript` tests covering
both good weather and bad weather scenarios (this also improves the good
weather scenarios for verification by covering the scenario where only
some of the checksums are valid and only those are being validated for
both workflow and job verification). Additionally, local functionality
is unit tested in places where unit tests already exist.