Continuously validate ghasum checksums at the job level
Remove the GitHub Actions workflow that continuously validates the
ghasum checksums for all Actions in this project in favor of a per-job
step that validates the ghasum checksums for that job only. This reduces
makes jobs self-validating, which prevents potentially compromised jobs
from executing while allowing other jobs to keep running.

The implementation leverages GitHub Actions context values to determine
what job to validate, resulting in an identical step in all jobs. This
makes it a good proof of concept for a ghasum action that can be used in
any job.

This commit is a followup to <2effc08890ff0aaa80d46a390489693b3c3deddc>.