Test and document behavior of sumfiles with duplicate entries
Add tests that verify the `ghasum update` and `ghasum verify` commands
both reject sumfiles with duplicate (non-header) entries. Additionally
test that `ghasum update -force` fixes duplicate (non-header) entries.
To support this change some test re-organization was in order. The pre-
existing "invalid sumfile" tests were specifically testing against a
sumfile with a syntax error, hence these have been renamed (given the
new tests are also for an "invalid sumfile").

Additionally, the implementation for sumfile parsing has been updated to
improve the error messages produced, which are asserted in the new and
changed tests. Similarly, the `SPECIFICATION.md` is updated to make
explicit that duplicate sumfile entries should be rejected.

As an aside, this uncovered a off-by-one bug in the line number formula
in the implementation for sumfile parsing and corresponding incorrectly
implemented unit test - both of which are fixed here.