Correct `uses:` value parsing with respect to the at sign
Improve the internal `parseUses` function to better handle the `@` sign.
In particular, only one of `@` is allowed in the `uses:` value. Before,
when multiple appeared in the value, it could be parsed incorrectly.
Given such values are invalid, an incorrectly parsed value would result
in a downstream error.

This problem was detected by a flaky test failure in the `constructive`
property test for the `parseUses` function.