Add `-offline` verification support
Update the `ghasum verify` command to accept the `-offline` flag and, if
provided, don't download repositories from the internet and rely on the
cache exclusively. If a repository is missing from the cache this causes
an error, avoiding a scenario where a missing repository is (silently)
not verified against the sumfile.

This flag is also applied in CI to the "Verify action checksums" steps
to ensure it verifies the checksums against the actions that will be
used rather a downloaded action (that won't be used).

The `-offline` flag is not extensively tested because all (testdata)
tests are already offline. Instead, only bad weather tests for the flag
have been added.