Correct flaky test for `parseUses`
Fix a flaky test for the `parseUses` function which could fail if the
randomly generated `ref` value contained an at sign. This would result
in a constructed `uses:` value with multiple at signs, which is not
valid (and would thus error unexpectedly).

This is a followup to 8588dffc1de7aa9d0a250e226915d688bd02395c