Correct checksum verification for `Semgrep/Semgrep` job
Update the "Verify action checksums" step for the "Semgrep" job in the
"Semgrep" workflow. The cache location for this job is different from
other jobs because it's running in a container (see `container:`
directive at the job level).

IMPORTANT: This is a debug commit and should not be merge as is.