Test and document behavior of sumfiles with duplicate headers
Add tests that verify the `ghasum update` and `ghasum verify` commands
both reject sumfiles with duplicate headers. To support this change some
test re-organization was in order. The existing "Sumfile with ..." tests
were specifically testing against problems with (non-header) entries but
now there are (separate) tests for problems with headers and entries.
Also, duplicate headers are unit tested for regardless of the sumfile
version.

The behavior of `ghasum update -force` is not covered here because there
is currently no defined behavior for what `ghasum update` should do with
headers.

Additionally, the implementation for sumfile parsing has been updated to
improve the error messages produced, which are asserted in the new and
changed tests. Similarly, the `SPECIFICATION.md` is updated to make
explicit that duplicate headers in a sumfile should be rejected.

As an aside, this (also) uncovered a off-by-one bug in the line number
formula in the implementation for sumfile header parsing which is fixed
here as well.