Set up auto-update workflow for gha.sum
Create a GitHub Actions workflow that automatically updates the
checksums in gha.sum for Pull Requests by Dependabot that update a
GitHub Action (leveraging `branches: dependabot/github_actions/**`).

This workflow utilizes a bot to be able to push the changes as well as
cause the created commit to trigger CI workflows. The secret it utilizes
must be configured for Dependabot (go to repository Settings > Secrets
and variables > Dependabot). The bot also requires permissions to "Read
and write" for the "Repository permissions" category called "Workflows".

The `if: ${{ github.actor == 'dependabot[bot]' }}` condition ensures the
workflow isn't (re-)run for the commit it creates. Besides avoiding a
potential infinite loop, it also avoids an error due to the (Dependabot
scoped) secrets not being available.

This change introduces two new GitHub Actions dependencies. One is used
to generate an access token from a GitHub app  The other is used to
simplify the creation and pushing of a commit to the Pull Request
branch.