Implement basic cache eviction
Add a simple and basic cache eviction strategy in order to avoid growing
the cache indefinitely and to avoid validating against cache entries
indefinitely (which may lead to false negatives). The current eviction
strategy is to remove anything older than 5 days.

Cache eviction happens in three scenarios. It happens when updating or
verifying (unless the `-no-evict` flag is used, more on that below), or
when explicitly using `ghasum cache evict`.

Eviction is done at the repository level. That is, a cached repository
will be evicted if it was cloned more than 5 days ago. Note that repos
are cloned and stored on a per version basis, so the eviction should
happen on, e.g., `actions/checkout/v5.0.0` (and not `actions/checkout`).

Cache eviction can be disabled when updating or verifying to support use
cases where eviction isn't desired. One example is when there is no
internet connection. Another example is when validating in a CI pipeline
(because we really want to validate with the runner put there, even if
its timestamp indicates it's old).