Update the Background section in the README
Update the README with an improved Background section following a recent
GH blog post: <https://github.blog/2024-04-29-highlights-from-git-2-45>.
Based on it, this adds the SHAmbles attack (in addition to SHAttered)
and Git's hash function transition. I also took the liberty of rewriting
the sentence on GitHub's protections which were both factually incorrect
(hash functions aren't probabilistic) and misleading (while the
mitigations *are* probabilistic, the security parameters is 2^-90, so
the use of that term may be misleading).