Make `ghasum update -force` update existing checksums that are incorrect
Update the implementation of the `ghasum update` command to update any
existing sums that are incorrect IF the `-force` flag is used. Here,
"existing sums" refers to entries in the sumfile for which the whole ID
matches before and after the update. This supports updating such entries
natively (i.e. without having to manually edit the sumfile), which may
be necessary if the incident has been reviewed and deemed acceptable.

The CLI help message as well as the specification text has been updated
accordingly.